requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pytest
pytest-asyncio
httpx
pillow
ultralytics
python-dotenv